# attempt to optimise vectored write